backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyjwt>=2.8.0
python-multipart>=0.0.6
passlib>=1.7.4
bcrypt>=4.0.0
email-validator>=2.0.0
python-dateutil>=2.8.0
fastapi-cors>=0.0.6
python-json-logger>=2.0.0
